replace defcustom "s3ed-prefix" with "s3ed-bucket-list"
A single prefix was too restrictive.  This change introduces
`s3ed-bucket-list', a list of buckets that one has access to (default empty
list).

Use M-n/p to access contents of the list.

The implementation is not great.  Elements of the list are presented for
`current-s3-base-path' in `s3ed-find-file' before the `s3ed-completing-read'.
This is to avoid `s3ed-completing-read' from erroring out when an inaccessible
bucket is in the minibuffer.